+++ ITechArtPizzaDelivery.Domain/Models/CartPizza.cs
@@ -8,9 +8,9 @@ namespace ITechArtPizzaDelivery.Domain.Models
 {
     public class CartPizza
     {
-        public long CartId { get; set; }
+        public int CartId { get; set; }
         public Cart Cart { get; set; }
-        public long PizzaId { get; set; }
+        public int PizzaId { get; set; }
         public Pizza Pizza { get; set; }
         public int Count { get; set; }
     }
